+++ app/happy_box.rb
@@ -1,0 +1,8 @@
+class HappyBox < Liza::HappyBox
+
+  axo do
+    # Set up your axo panel per the DSL in http://guides.lizarb.com/panels/axo.html
+
+  end
+
+end