+++ lib/happy_system.rb
@@ -1,0 +1,12 @@
+# frozen_string_literal: true
+
+class HappySystem < Liza::System
+  class Error < StandardError; end
+  # Your code goes here...
+
+  #
+
+  set :log_color, :magenta
+
+  #
+end