+++ lib/happy_system/happy/panels/axo_panel.rb
@@ -1,0 +1,5 @@
+class HappySystem
+  class AxoPanel < Liza::Panel
+
+  end
+end