+++ CHANGELOG.md
@@ -7,3 +7,4 @@
 - Liza::Unit has class/instance methods log/log_color/log_level/log?/log_level?
 - Liza::System, Liza::Box, Liza::Panel, Liza::Controller
 - DevSystem, Liza::DevBox, Liza::Command, AppCommand, and others
+- HappySystem, Liza::HappyBox, Liza::Axo, Axo, and others